separated styles the components ui
