+++ ui/about/styles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const Wrapper = styled.section`
+  > * {
+    margin-bottom: 19rem;
+  }
+`
