+++ ui/index.js
@@ -4,21 +4,21 @@ export * from './global-styles'
 export * from './media'
 
 // Components
-export * from './icons/icons'
-export * from './container/container'
-export * from './button/button'
-export * from './head/head'
-export * from './header/header'
-export * from './hero/hero'
-export * from './about/about'
-export * from './video/video'
-export * from './empty-video/empty-video'
-export * from './playlist/playlist'
-export * from './issue-list/issue-list'
-export * from './footer/footer'
-export * from './link/link'
-export * from './bold-text/bold-text'
-export * from './text-box/text-box'
-export * from './issue-card/issue-card'
-export * from './cta-box/cta-box'
-export * from './follow-up/follow-up'
+export * from './icons'
+export * from './container'
+export * from './button'
+export * from './head'
+export * from './header'
+export * from './hero'
+export * from './about'
+export * from './video'
+export * from './empty-video'
+export * from './playlist'
+export * from './issue-list'
+export * from './footer'
+export * from './link'
+export * from './bold-text'
+export * from './text-box'
+export * from './issue-card'
+export * from './cta-box'
+export * from './follow-up'
