+++ ui/issue-list/index.js
@@ -1,0 +1,22 @@
+import { TextBox, IssueCard, EmptyIssuesIcon } from 'ui'
+import { issueList } from 'resources/content'
+
+import { Wrapper, List, Empty, Text } from './styles'
+
+export const IssueList = () => {
+  return (
+    <Wrapper id='contribua'>
+      <TextBox title={issueList.title} text={issueList.text} />
+      <List>
+        {issueList.length && issueList.items.map(({ id, ...card }) => (
+          <IssueCard key={id} {...card} />
+        ))}
+      </List>
+      {!issueList.length &&
+        <Empty>
+          <EmptyIssuesIcon />
+          <Text>Em breve as issues estarão disponíveis aqui.</Text>
+        </Empty>}
+    </Wrapper>
+  )
+}
